+++ buildSrc/src/main/java/net/lopymine/mossy/manager/MossyDependenciesManager.java
@@ -1,15 +1,15 @@
 package net.lopymine.mossy.manager;
 
-import dev.kikugie.stonecutter.StonecutterBuild;
 import lombok.experimental.ExtensionMethod;
 import org.gradle.api.*;
 import org.gradle.api.artifacts.dsl.DependencyHandler;
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository;
 
 import net.lopymine.mossy.*;
-import net.lopymine.mossy.extension.MossyDependenciesExtension;
+import net.lopymine.mossy.extension.*;
+import net.lopymine.mossy.extension.MossyAdditionalDependencies.AdditionalDependencyOverride;
 
-import java.util.Map;
+import java.util.*;
 import java.util.function.Consumer;
 import org.jetbrains.annotations.NotNull;
 
@@ -32,18 +32,50 @@ public class MossyDependenciesManager {
 		dependencies.add("annotationProcessor", "org.projectlombok:lombok:%s".formatted(lombok));
 
 		Map<String, String> properties = project.getMossyProperties("dep");
-		String yacl = properties.remove("yacl");
+		MossyAdditionalDependencies additional = extension.getAdditional();
+		additional.disable("yacl");
 
-		if (yacl != null) {
-			dependencies.add("modImplementation", "dev.isxander:yet-another-config-lib:%s".formatted(yacl));
-		}
+		Map<String, AdditionalDependencyOverride> overrides = additional.getOverrides();
+		Set<String> disabled = additional.getDisabled();
+		properties.forEach((modId, version) -> {
+			if (disabled.contains(modId)) {
+				//System.out.println("Disabling auto-dependency for %s !".formatted(modId));
+				return;
+			}
 
-		properties.forEach((key, value) -> {
-			dependencies.add("modImplementation", "maven.modrinth:%s:%s".formatted(key, value));
+			if (version.equals("unknown")) {
+				return;
+			}
+
+			AdditionalDependencyOverride override = overrides.get(modId);
+			String configurationName = override != null ? override.configurationName() : "modImplementation";
+			dependencies.add(configurationName, "maven.modrinth:%s:%s".formatted(modId, version));
 		});
+
+		project.getConfigurations().forEach((configuration) -> {
+			configuration.resolutionStrategy((strategy) -> {
+				strategy.force("com.twelvemonkeys.common:common-io:3.10.0");
+				strategy.force("com.twelvemonkeys.common:common-lang:3.10.0");
+				strategy.force("com.twelvemonkeys.common:common-image:3.10.0");
+				strategy.force("com.twelvemonkeys.imageio:imageio-metadata:3.10.0");
+				strategy.force("com.twelvemonkeys.imageio:imageio-webp:3.10.0");
+				strategy.force("com.twelvemonkeys.imageio:imageio-core:3.10.0");
+			});
+		});
+
+		String yaclVersion = properties.get("yacl");
+		if (yaclVersion != null && !yaclVersion.equals("unknown")) {
+			Set<String> oldMavenVersions = Set.of("1.19.4", "1.20", "1.20.2", "1.20.3");
+			if (oldMavenVersions.contains(minecraft)) {
+				dependencies.add("modImplementation", "dev.isxander.yacl:yet-another-config-lib-fabric:%s".formatted(MossyPlugin.substringBeforeLast(yaclVersion, "-")));
+			} else {
+				dependencies.add("modImplementation", "dev.isxander:yet-another-config-lib:%s".formatted(yaclVersion));
+			}
+		}
 	}
 
 	private static void addRepositories(Project project) {
+		project.getRepositories().mavenCentral();
 		addRepository(project, "Quilt", "https://maven.quiltmc.org/repository/release/");
 		addRepository(project, "Sonatype", "https://oss.sonatype.org/content/repositories/snapshots/");
 		addRepository(project, "Terraformers", "https://maven.terraformersmc.com/");
@@ -54,6 +86,7 @@ public class MossyDependenciesManager {
 				descriptor.includeGroup("maven.modrinth");
 			});
 		});
+
 	}
 
 	private static void addRepository(Project project, String name, String url) {