+++ src/main/java/net/lopymine/te/TransparentEntities.java
@@ -9,7 +9,7 @@ public class TransparentEntities implements ModInitializer {
 
 	public static final String MOD_NAME = /*$ mod_name*/ "Transparent Entities";
 	public static final String MOD_ID = /*$ mod_id*/ "transparent-entities";
-	public static final String YACL_DEPEND_VERSION = /*$ yacl*/ "3.6.2+1.21.4-fabric";
+	public static final String YACL_DEPEND_VERSION = /*$ yacl*/ "3.7.1+1.21.6-fabric";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_NAME);
 
 	public static Identifier id(String path) {