+++ src/main/java/net/lopymine/te/mixin/entity/EntityRenderDispatcherShadowMixin.java
@@ -20,7 +20,7 @@ import net.minecraft.client.render.entity.state.EntityRenderState;
 public class EntityRenderDispatcherShadowMixin {
 
 	//? if >=1.21.5 {
-	/*@WrapOperation(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/entity/EntityRenderer;getShadowOpacity(Lnet/minecraft/client/render/entity/state/EntityRenderState;)F"), method = "render(Lnet/minecraft/client/render/entity/state/EntityRenderState;DDDLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/client/render/entity/EntityRenderer;)V")
+	@WrapOperation(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/entity/EntityRenderer;getShadowOpacity(Lnet/minecraft/client/render/entity/state/EntityRenderState;)F"), method = "render(Lnet/minecraft/client/render/entity/state/EntityRenderState;DDDLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/client/render/entity/EntityRenderer;)V")
 	private float generated222(EntityRenderer instance, EntityRenderState entityRenderState, Operation<Float> original) {
 		float call = original.call(instance, entityRenderState);
 		Entity entity = EntityCaptures.MAIN.getEntity();
@@ -30,8 +30,8 @@ public class EntityRenderDispatcherShadowMixin {
 		int originalColor = ArgbUtils.getArgb((int) (call * 255F), 255, 255, 255);
 		return ArgbUtils.getAlpha(TransparencyManager.getTranslucentArgb(entity, originalColor)) / 255F;
 	}
-	*///?} elif >=1.21.4 {
-	@WrapOperation(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/entity/EntityRenderer;getShadowOpacity(Lnet/minecraft/client/render/entity/state/EntityRenderState;)F"), method = "render(Lnet/minecraft/entity/Entity;DDDFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/client/render/entity/EntityRenderer;)V")
+	//?} elif >=1.21.4 {
+	/*@WrapOperation(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/entity/EntityRenderer;getShadowOpacity(Lnet/minecraft/client/render/entity/state/EntityRenderState;)F"), method = "render(Lnet/minecraft/entity/Entity;DDDFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/client/render/entity/EntityRenderer;)V")
 	private float generated(EntityRenderer instance, EntityRenderState entityRenderState, Operation<Float> original, @Local(argsOnly = true) Entity entity) {
 		float call = original.call(instance, entityRenderState);
 		if (!TransparencyManager.canRenderTransparencyShadow(entity)) {
@@ -40,7 +40,7 @@ public class EntityRenderDispatcherShadowMixin {
 		int originalColor = ArgbUtils.getArgb((int) (call * 255F), 255, 255, 255);
 		return ArgbUtils.getAlpha(TransparencyManager.getTranslucentArgb(entity, originalColor)) / 255F;
 	}
-	//?} elif >=1.21.3 {
+	*///?} elif >=1.21.3 {
 	/*@ModifyExpressionValue(at = @At(value = "FIELD", target = "Lnet/minecraft/client/render/entity/EntityRenderer;shadowOpacity:F"), method = "render(Lnet/minecraft/entity/Entity;DDDFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/client/render/entity/EntityRenderer;)V")
 	private float generated(float originalAlpha, @Local(argsOnly = true) Entity entity) {
 		if (!TransparencyManager.canRenderTransparencyShadow(entity)) {