+++ src/main/java/net/lopymine/te/mixin/item/ItemRendererMixin.java
@@ -29,16 +29,16 @@ public class ItemRendererMixin {
 		return TransparencyLayers.getLayer(() -> original.call(stack));
 	}
 	*///?} elif <=1.21.4 {
-	@ModifyVariable(at = @At("HEAD"), method = "renderItem(Lnet/minecraft/item/ModelTransformationMode;Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;II[ILnet/minecraft/client/render/model/BakedModel;Lnet/minecraft/client/render/RenderLayer;Lnet/minecraft/client/render/item/ItemRenderState$Glint;)V", argsOnly = true, index = 7)
+	/*@ModifyVariable(at = @At("HEAD"), method = "renderItem(Lnet/minecraft/item/ModelTransformationMode;Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;II[ILnet/minecraft/client/render/model/BakedModel;Lnet/minecraft/client/render/RenderLayer;Lnet/minecraft/client/render/item/ItemRenderState$Glint;)V", argsOnly = true, index = 7)
 	private static RenderLayer modifyLayer(RenderLayer value) {
 		return TransparencyLayers.getItemLayer(value);
 	}
-	//?} else {
-	/*@ModifyVariable(at = @At("HEAD"), method = "renderItem(Lnet/minecraft/item/ItemDisplayContext;Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;II[ILjava/util/List;Lnet/minecraft/client/render/RenderLayer;Lnet/minecraft/client/render/item/ItemRenderState$Glint;)V", argsOnly = true, index = 7)
+	*///?} else {
+	@ModifyVariable(at = @At("HEAD"), method = "renderItem(Lnet/minecraft/item/ItemDisplayContext;Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;II[ILjava/util/List;Lnet/minecraft/client/render/RenderLayer;Lnet/minecraft/client/render/item/ItemRenderState$Glint;)V", argsOnly = true, index = 7)
 	private static RenderLayer modifyLayer(RenderLayer value) {
 		return TransparencyLayers.getItemLayer(value);
 	}
-	*///?}
+	//?}
 
 	//? if <=1.21.1 {
 	/*@WrapOperation(at = @At(value = "INVOKE", target = "Lnet/minecraft/util/math/ColorHelper$Argb;getAlpha(I)I"), method = "renderBakedItemQuads")