+++ src/main/java/net/lopymine/te/render/TransparencyLayers.java
@@ -3,11 +3,12 @@ package net.lopymine.te.render;
 import net.minecraft.client.render.*;
 import net.minecraft.client.render.RenderLayer.MultiPhaseParameters;
 //? if <=1.21.4 {
-import net.minecraft.client.render.RenderPhase.Texture;
+/*import net.minecraft.client.render.RenderPhase.Texture;
 import net.minecraft.client.render.VertexFormat.DrawMode;
-//?} else {
-/*import net.minecraft.client.gl.RenderPipelines;
-*///?}
+*///?} else {
+import net.minecraft.client.gl.RenderPipelines;
+//?}
+import net.minecraft.client.render.RenderPhase.Texture;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.*;
 
@@ -28,16 +29,16 @@ public class TransparencyLayers {
 	private static final /*? >=1.21.2 {*/ TriState /*?} else {*/ /*boolean*/ /*?}*/ idk = /*? >=1.21.2 {*/ TriState.FALSE /*?} else {*/ /*false*/ /*?}*/;
 
 	//? if <=1.21.4 {
-	private static final Function<Identifier, RenderLayer> ITEM_ENTITY_TRANSLUCENT_NO_CULL = Util.memoize((texture) -> {
+	/*private static final Function<Identifier, RenderLayer> ITEM_ENTITY_TRANSLUCENT_NO_CULL = Util.memoize((texture) -> {
 		MultiPhaseParameters multiPhaseParameters = MultiPhaseParameters.builder().program(RenderLayer.ITEM_ENTITY_TRANSLUCENT_CULL_PROGRAM).texture(new Texture(texture, idk, false)).transparency(RenderLayer.TRANSLUCENT_TRANSPARENCY).cull(RenderLayer.DISABLE_CULLING).target(RenderLayer.ITEM_ENTITY_TARGET).lightmap(RenderLayer.ENABLE_LIGHTMAP).overlay(RenderLayer.ENABLE_OVERLAY_COLOR).writeMaskState(RenderLayer.ALL_MASK).build(true);
 		return RenderLayer.of("item_entity_translucent_no_cull", VertexFormats.POSITION_COLOR_TEXTURE_OVERLAY_LIGHT_NORMAL, DrawMode.QUADS, 1536, true, true, multiPhaseParameters);
 	});
-	//?} else {
-	/*private static final Function<Identifier, RenderLayer> ITEM_ENTITY_TRANSLUCENT_NO_CULL = Util.memoize((texture) -> {
-		MultiPhaseParameters multiPhaseParameters = RenderLayer.MultiPhaseParameters.builder().texture(new RenderPhase.Texture(texture, TriState.FALSE, false)).target(RenderLayer.ITEM_ENTITY_TARGET).lightmap(RenderLayer.ENABLE_LIGHTMAP).overlay(RenderLayer.ENABLE_OVERLAY_COLOR).build(true);
+	*///?} else {
+	private static final Function<Identifier, RenderLayer> ITEM_ENTITY_TRANSLUCENT_NO_CULL = Util.memoize((texture) -> {
+		MultiPhaseParameters multiPhaseParameters = MultiPhaseParameters.builder().texture(new Texture(texture, /*? if <=1.21.5 {*//*TriState.FALSE,*//*?}*/ false)).target(RenderLayer.ITEM_ENTITY_TARGET).lightmap(RenderLayer.ENABLE_LIGHTMAP).overlay(RenderLayer.ENABLE_OVERLAY_COLOR).build(true);
 		return RenderLayer.of("item_entity_translucent_cull", 1536, true, true, TransparencyRenderPipelines.RENDER_TYPE_ITEM_ENTITY_TRANSLUCENT_NO_CULL, multiPhaseParameters);
 	});
-	*///?}
+	//?}
 
 	//? if >=1.21.2 {
 	public static RenderLayer getLayerNoCull(LayerType layerType, Identifier texture, Supplier<RenderLayer> original) {