+++ src/main/java/net/lopymine/te/render/TransparencyRenderPipelines.java
@@ -1,21 +1,25 @@
 package net.lopymine.te.render;
 
 //? if >=1.21.5 {
-/*import net.minecraft.client.gl.RenderPipelines;
+import net.minecraft.client.gl.RenderPipelines;
 import net.minecraft.client.render.VertexFormats;
 
 import com.mojang.blaze3d.pipeline.*;
 import com.mojang.blaze3d.vertex.VertexFormat.DrawMode;
+//?}
 
-import static net.minecraft.client.gl.RenderPipelines.MATRICES_COLOR_FOG_LIGHT_DIR_SNIPPET;
-*///?}
 
 public class TransparencyRenderPipelines {
 
 	//? if >=1.21.5 {
-	/*public static final RenderPipeline RENDER_TYPE_ITEM_ENTITY_TRANSLUCENT_NO_CULL =
-			RenderPipeline.builder(MATRICES_COLOR_FOG_LIGHT_DIR_SNIPPET)
-					.withLocation("pipeline/item_entity_translucent_cull")
+	public static final RenderPipeline RENDER_TYPE_ITEM_ENTITY_TRANSLUCENT_NO_CULL =
+			RenderPipeline.builder(
+					//? if >=1.21.6 {
+					net.minecraft.client.gl.RenderPipelines.TRANSFORMS_PROJECTION_FOG_LIGHTING_SNIPPET
+					//?} elif =1.21.5 {
+					/*net.minecraft.client.gl.RenderPipelines.MATRICES_COLOR_FOG_LIGHT_DIR_SNIPPET
+	                *///?}
+					).withLocation("pipeline/item_entity_translucent_cull")
 					.withVertexShader("core/rendertype_item_entity_translucent_cull")
 					.withFragmentShader("core/rendertype_item_entity_translucent_cull")
 					.withSampler("Sampler0")
@@ -28,10 +32,11 @@ public class TransparencyRenderPipelines {
 	public static void register() {
 		RenderPipelines.register(RENDER_TYPE_ITEM_ENTITY_TRANSLUCENT_NO_CULL);
 	}
-	*///?} else {
-	public static void register() {
+	//?} else {
+	/*public static void register() {
 
 	}
-	//?}
+	*///?}
+
 
 }